# docs/api/openapi.yaml
openapi: 3.1.1
info:
  title: Books Memo API
  version: 1.0.0
  summary: Personal reading tracker API
  description: >-
    Books Memo 向けの REST API。書誌原本(book_master)と個人利用情報(user_book)を分離し、
    ISBN 取り込み、タグ管理、読書状態更新を提供する。
servers:
  - url: https://api.example.com/api/v1
    description: production
  - url: http://localhost:8080/api/v1
    description: local
security:
  - bearerAuth: []
tags:
  - name: Auth
  - name: Books
  - name: ISBN
  - name: Tags
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout
      responses:
        '204':
          description: Logout success
  /me:
    get:
      tags: [Auth]
      summary: Get current user
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
  /books:
    get:
      tags: [Books]
      summary: Search books
      parameters:
        - in: query
          name: keyword
          schema: { type: string, maxLength: 100 }
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/BookStatus'
        - in: query
          name: tagId
          schema: { type: integer, format: int64 }
        - in: query
          name: favorite
          schema: { type: boolean }
        - in: query
          name: sort
          schema:
            type: string
            enum: [updatedAtDesc, createdAtDesc, titleAsc, ratingDesc, publishedDateDesc]
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: size
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: Paged list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookListResponse'
    post:
      tags: [Books]
      summary: Create manual book entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserBookDetail'
        '409':
          description: Duplicate active record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /books/external-search:
    get:
      tags: [Books]
      summary: Search external book metadata
      parameters:
        - in: query
          name: query
          required: true
          schema: { type: string, minLength: 1 }
        - in: query
          name: type
          schema:
            $ref: '#/components/schemas/ExternalBookSearchType'
        - in: query
          name: maxResults
          schema: { type: integer, minimum: 1, maximum: 40, default: 10 }
      responses:
        '200':
          description: External book candidates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExternalBookSearchResponse'
  /books/import-by-isbn:
    post:
      tags: [Books, ISBN]
      summary: Import book by ISBN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportByIsbnRequest'
      responses:
        '201':
          description: Imported
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserBookDetail'
        '404':
          description: ISBN not found in providers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /books/{userBookId}:
    get:
      tags: [Books]
      summary: Get detail
      parameters:
        - $ref: '#/components/parameters/UserBookId'
      responses:
        '200':
          description: Detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserBookDetail'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    patch:
      tags: [Books]
      summary: Update book info
      parameters:
        - $ref: '#/components/parameters/UserBookId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBookRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserBookDetail'
    delete:
      tags: [Books]
      summary: Logical delete
      parameters:
        - $ref: '#/components/parameters/UserBookId'
      responses:
        '204':
          description: Deleted
  /books/{userBookId}/status:
    patch:
      tags: [Books]
      summary: Update reading status only
      parameters:
        - $ref: '#/components/parameters/UserBookId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStatusRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserBookDetail'
  /tags:
    get:
      tags: [Tags]
      summary: List tags
      responses:
        '200':
          description: Tag list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tag'
    post:
      tags: [Tags]
      summary: Create tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTagRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
  /tags/{tagId}:
    patch:
      tags: [Tags]
      summary: Update tag
      parameters:
        - $ref: '#/components/parameters/TagId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTagRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
    delete:
      tags: [Tags]
      summary: Delete tag
      parameters:
        - $ref: '#/components/parameters/TagId'
      responses:
        '204':
          description: Deleted
  /isbn/{isbn}:
    get:
      tags: [ISBN]
      summary: Lookup ISBN candidates without creating user_book
      parameters:
        - in: path
          name: isbn
          required: true
          schema:
            type: string
            pattern: '^[0-9X-]{10,17}$'
      responses:
        '200':
          description: Candidate data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IsbnLookupResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserBookId:
      in: path
      name: userBookId
      required: true
      schema: { type: integer, format: int64 }
    TagId:
      in: path
      name: tagId
      required: true
      schema: { type: integer, format: int64 }
  schemas:
    BookStatus:
      type: string
      enum: [WISHLIST, PURCHASED, READING, FINISHED, ON_HOLD, DROPPED, TSUNDOKU]
    ExternalBookSearchType:
      type: string
      enum: [KEYWORD, TITLE, AUTHOR, ISBN]
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8, maxLength: 72 }
    LoginResponse:
      type: object
      required: [accessToken, tokenType, expiresIn, user]
      properties:
        accessToken: { type: string }
        tokenType: { type: string, example: Bearer }
        expiresIn: { type: integer, example: 3600 }
        user:
          $ref: '#/components/schemas/MeResponse'
    MeResponse:
      type: object
      required: [id, email, displayName, role]
      properties:
        id: { type: integer, format: int64 }
        email: { type: string, format: email }
        displayName: { type: string }
        role: { type: string, enum: [USER, ADMIN] }
    BookMasterSummary:
      type: object
      required: [title, sourcePrimary]
      properties:
        id: { type: integer, format: int64 }
        isbn13: { type: string, nullable: true }
        isbn10: { type: string, nullable: true }
        title: { type: string }
        subtitle: { type: string, nullable: true }
        authors: { type: array, items: { type: string } }
        publisher: { type: string, nullable: true }
        publishedDate: { type: string, format: date, nullable: true }
        thumbnailUrl: { type: string, nullable: true }
        sourcePrimary: { type: string, enum: [OPENBD, GOOGLE_BOOKS, MANUAL] }
    Tag:
      type: object
      required: [id, name, sortOrder]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string, maxLength: 50 }
        colorHex: { type: string, pattern: '^#[0-9A-Fa-f]{6}$', nullable: true }
        sortOrder: { type: integer }
    BookListItem:
      type: object
      required: [id, status, favoriteFlag, bookMaster]
      properties:
        id: { type: integer, format: int64 }
        status: { $ref: '#/components/schemas/BookStatus' }
        rating: { type: number, format: double, nullable: true }
        favoriteFlag: { type: boolean }
        purchaseDate: { type: string, format: date, nullable: true }
        startDate: { type: string, format: date, nullable: true }
        finishDate: { type: string, format: date, nullable: true }
        updatedAt: { type: string, format: date-time }
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        bookMaster:
          $ref: '#/components/schemas/BookMasterSummary'
    PaginationMeta:
      type: object
      required: [page, size, totalItems, totalPages]
      properties:
        page: { type: integer }
        size: { type: integer }
        totalItems: { type: integer, format: int64 }
        totalPages: { type: integer }
    BookListResponse:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/BookListItem'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    UserBookDetail:
      allOf:
        - $ref: '#/components/schemas/BookListItem'
        - type: object
          properties:
            memo: { type: string, nullable: true }
            locationNote: { type: string, nullable: true }
            createdAt: { type: string, format: date-time }
            deletedAt: { type: string, format: date-time, nullable: true }
    CreateBookRequest:
      type: object
      required: [title, status]
      properties:
        isbn13: { type: string, nullable: true }
        isbn10: { type: string, nullable: true }
        title: { type: string, maxLength: 500 }
        subtitle: { type: string, nullable: true }
        authors: { type: array, items: { type: string } }
        publisher: { type: string, nullable: true }
        publishedDate: { type: string, format: date, nullable: true }
        description: { type: string, nullable: true }
        thumbnailUrl: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/BookStatus' }
        rating: { type: number, format: double, nullable: true }
        favoriteFlag: { type: boolean, default: false }
        memo: { type: string, nullable: true }
        tagIds:
          type: array
          items: { type: integer, format: int64 }
    ImportByIsbnRequest:
      type: object
      required: [isbn, status]
      properties:
        isbn: { type: string }
        status: { $ref: '#/components/schemas/BookStatus' }
        memo: { type: string, nullable: true }
        tagIds:
          type: array
          items: { type: integer, format: int64 }
    UpdateBookRequest:
      type: object
      properties:
        status: { $ref: '#/components/schemas/BookStatus' }
        rating: { type: number, format: double, nullable: true }
        favoriteFlag: { type: boolean }
        purchaseDate: { type: string, format: date, nullable: true }
        startDate: { type: string, format: date, nullable: true }
        finishDate: { type: string, format: date, nullable: true }
        memo: { type: string, nullable: true }
        locationNote: { type: string, nullable: true }
        tagIds:
          type: array
          items: { type: integer, format: int64 }
    UpdateStatusRequest:
      type: object
      required: [status]
      properties:
        status: { $ref: '#/components/schemas/BookStatus' }
    CreateTagRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 50 }
        colorHex: { type: string, pattern: '^#[0-9A-Fa-f]{6}$', nullable: true }
        sortOrder: { type: integer, default: 0 }
    UpdateTagRequest:
      type: object
      properties:
        name: { type: string, maxLength: 50 }
        colorHex: { type: string, pattern: '^#[0-9A-Fa-f]{6}$', nullable: true }
        sortOrder: { type: integer }
    IsbnLookupCandidate:
      type: object
      required: [isbn13, title, sourceName]
      properties:
        isbn13: { type: string }
        isbn10: { type: string, nullable: true }
        title: { type: string }
        subtitle: { type: string, nullable: true }
        authors: { type: array, items: { type: string } }
        publisher: { type: string, nullable: true }
        publishedDate: { type: string, format: date, nullable: true }
        thumbnailUrl: { type: string, nullable: true }
        description: { type: string, nullable: true }
        sourceName: { type: string, enum: [OPENBD, GOOGLE_BOOKS, MANUAL] }
        cacheHit: { type: boolean }
    IsbnLookupResponse:
      type: object
      required: [queryIsbn, candidates]
      properties:
        queryIsbn: { type: string }
        candidates:
          type: array
          items:
            $ref: '#/components/schemas/IsbnLookupCandidate'
    ExternalBookSearchCandidate:
      type: object
      required: [title, sourceName]
      properties:
        isbn13: { type: string, nullable: true }
        isbn10: { type: string, nullable: true }
        title: { type: string }
        subtitle: { type: string, nullable: true }
        authors: { type: array, items: { type: string } }
        publisher: { type: string, nullable: true }
        publishedDate: { type: string, format: date, nullable: true }
        thumbnailUrl: { type: string, nullable: true }
        description: { type: string, nullable: true }
        sourceName: { type: string, enum: [GOOGLE_BOOKS] }
    ExternalBookSearchResponse:
      type: object
      required: [query, type, candidates]
      properties:
        query: { type: string }
        type:
          $ref: '#/components/schemas/ExternalBookSearchType'
        candidates:
          type: array
          items:
            $ref: '#/components/schemas/ExternalBookSearchCandidate'
    ApiError:
      type: object
      required: [code, message, requestId]
      properties:
        code: { type: string, example: BOOK_DUPLICATE }
        message: { type: string }
        requestId: { type: string }
        details:
          type: array
          items: { type: string }